backend/requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
PyMySQL==1.0.3
pytest==7.2.0  # Add pytest here
